requirements.txt
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-allauth==0.38.0
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-ranged-response==0.2.0
django-reversion==3.0.2
django-simple-captcha==0.5.9
django-user-accounts==2.1.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.7
itsdangerous==1.1.0
jdcal==1.4
mysqlclient==1.3.14
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
pycodestyle==2.4.0
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0
